Update new excersices and Utils.py
